feat: Import User interface in post.d.ts
Import the User interface from the user module in the post.d.ts file to use it in the Post interface.

Refactor: Update import path in nav-user.tsx and nav-user.tsx

Update the import path for the SignOut action in the nav-user.tsx and nav-user.tsx files to use the new location in the auth.actions module.

Refactor: Update import path in page.tsx and [slug]/page.tsx

Update the import path for the getUsers and getPostBySlug functions in the page.tsx and [slug]/page.tsx files to use the new location in the user.actions and post.actions modules respectively.

Refactor: Update default variant in toast.tsx

Update the default variant in the toast.tsx file to use the "text-teal-700" color for the text instead of "text-foreground".

Refactor: Update className in layout.tsx

Update the className in the layout.tsx file to add the "selection:bg-teal-100" and "selection:text-teal-900" styles for the text selection.

Refactor: Update import path in recent-users.tsx and post-list.tsx

Update the import path for the getUsers function in the recent-users.tsx and post-list.tsx files to use the new location in the user.actions and post.actions modules respectively.

Refactor: Update import path in login-form.tsx and register-form.tsx

Update the import path for the loginCredentials and registerCredentials actions in the login-form.tsx and register-form.tsx files to use the new location in the auth.actions module.

Refactor: Remove unused user.services module

Remove the unused user.services module as it is no longer needed.

Refactor: Remove unused services folder

Remove the unused services folder as it is no longer needed.

Refactor: Update import path in post-form.tsx

Update the import path for the createPost action in the post-form.tsx file to use the new location in the post.actions module.